+++ example_game/src/main.rs
@@ -196,6 +196,6 @@ fn panic_hook(info: &std::panic::PanicInfo) {
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen(start))]
 pub async fn run() {
     std::panic::set_hook(Box::new(panic_hook));
-    console_log::init_with_level(log::Level::Debug).expect("Couldn't initialize logger");
+    console_log::init_with_level(log::Level::Info).expect("Couldn't initialize logger");
     start().await;
 }
